// sdk/typescript/models/OrganizationDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type OrganizationDto = {
    readonly organizationId?: number;
    readonly name?: string | null;
    readonly description?: string | null;
    readonly url?: string | null;
    readonly phone?: string | null;
    readonly email?: string | null;
    readonly logoUrl?: string | null;
    readonly logoBase64?: string | null;
};