// sdk/typescript/models/OrganizationFormDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type OrganizationFormDto = {
    name: string;
    description?: string | null;
    url?: string | null;
    phone?: string | null;
    email?: string | null;
    logoUrl?: string | null;
    logoBase64?: string | null;
};